team-sisi-kwa-sisi/backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.6
djoser==2.1.0
django-cors-headers==3.14.0
cloudinary==1.30.0
stripe==5.0.0
twilio==6.66.0
django-environ==0.8.1
leaflet==0.1.0
django-filter==23.2.0
django-extensions==3.2.1
Pillow==10.0.0
django-redis==5.2.0